Implemented cue registry, set existing cues to be default. NOT YET USING
Cue registry lets us create multiple banks/groups of cues.

When a cue is called the active group is checked for an implementation of that cue. If found it's used. If not, fall back to the default implementation of that cue.

This saves us from re-defining common cues like Strobe, etc.

We are not yet setting any groups as active, so it's always falling back to default.