WIP: Enabled/disabled groups impact active groups
The DMX preview UI now respects the enabled state.
In-game the underlying system forces a fallback to the default group even if its disabled.

Need to remove that fallback but then include notifications to the user that a cue is being skipped because none are available. Otherwise the lack of output would be confusing.