Updates to DMX Preview and Simulation handling, TODO: Vary cue selection
Clarify cue simulation, disable test buttons if a specific cue group and cue effect haven't been selected.

Selecting a cue to simulate will restrict the active cue groups to that group to ensure the specific effect is selected by the cue handlers.

TODO: getCueImplementation will alway use the same set of cues based on the registration order.
We need to add logic so this can be changed, but not simply randomized on every request.