WIP: Added randomization logic for cue implementation selection
Primary and secondary cues can switch if they are re-used too much. I.e when a cue changes to a new one we reset the counter to 0. Each time the same cue is called increment. If it passes the limit, randomly pick a new implementation of that cue.

If we change the primary or secondary cue, reset the counter and randomly select the implementation.

Primary and secondary have different limits