Current cue state revisions, front-end display
Cue state is sent from the BE to FE for accurate state. This is broadcast at the app level and tracked in an atom for use in varying components.

This will let us show the basic cue information on status as well as DMX preview, etc.

FE shows the current Primary/Secondary cue.