Inverting yellow in Default, faster blackout, removeAll when set
Set used some logic to decide how to clear things before we had lifecycles in the cues. Now with lifecycle we know set will only be called the first time, letting us always clear the effects immediately before the new effect is set.
(Unless layering on top with add on a higher layer)