Fix for waiting transitions states not being included
We were losing the state of layers that were waiting on an event because it was only calculating the final colours for lights actively changing.

E.g. layer 0 is red. Layer 101 goes yellow -> transparent -> yellow. We lost the red on 0 because its state wasn't included in the final colour calculations since it was waiting and not actively changing.