Blocking nodal execution engine implemented
Node maps are now parsed and executed at runtime vs. precompiled. Allows for stage-specific effects to be passed to the sequencer with callback to continue traversing the node map.

This allows for conditions to be properly evaluated as the node is traversed. Eg. exit condition is delay 2000ms, followed by a IF conditional check that branches between true and false. The if condition will only resolve after the previous node's 2000ms exit delay and then the appropriate true/false condition is executed.